statistics/neo4j_etl/pyproject.toml
[project]
name = "neo4j_etl"
version = "0.1"
dependencies = [
   "python-dotenv",
   "neo4j==5.14.1",
   "retry==0.9.2"
]

[project.optional-dependencies]
dev = ["black", "flake8"]